api/requirements.txt
fastapi==0.110.1
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.9
bcrypt>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.1
cryptography>=42.0.8
pydicom>=2.4.0
pillow>=10.0.0
numpy>=1.26.0
